feat:返回值统一处理

+++ tomato/service/disease.py
@@ -6,32 +6,23 @@ class DiseaseService():
     def __init__(self, disease_repository: IDiseaseRepository) -> None:
         self.disease_repository: IDiseaseRepository = disease_repository
 
-    def get_by_tittle(self, tittle: str) -> Disease:
-        try:
-            return self.disease_repository.get_by_tittle(tittle)
-        except:
-            return -1
+    def get_by_title(self, title: str) -> Disease | None:
+        return self.disease_repository.get_by_title(title)
 
     def get_all(self) -> list[Disease]:
         return self.disease_repository.get_all()
 
-    def get_by_id(self, id: int) -> Disease:
-        try:
-            return self.disease_repository.get_by_id(id)
-        except:
-            return -1
+    def get_by_id(self, id: int) -> Disease | None:
+        return self.disease_repository.get_by_id(id)
 
     def create(self, disease: Disease) -> int:
-        try:
-            if self.disease_repository.get_by_tittle(disease.tittle) is not None:
-                return -1
-        except:
-            pass
-        self.disease_repository.create(disease)
-        return 0
+        return self.disease_repository.create(disease)
 
-    def update(self, disease: Disease):
-        self.disease_repository.update(disease)
+    def update(self, disease: Disease) -> int:
+        return self.disease_repository.update(disease)
 
-    def delete(self, id: int):
-        self.disease_repository.delete(id)
+
+    def delete(self, id: int) -> int:
+        if self.get_by_id(id) is None:
+            return -1
+        return self.disease_repository.delete(id)
